[#51] 펫시터 정보 및 리뷰 화면 연결 완료
