--- MungNolZa/app/src/main/java/kr/co/lion/mungnolza/ui/matching_petsitter/MatchingPetsitterActivity.kt
@@ -1,158 +1,0 @@
-package kr.co.lion.mungnolza.ui.matching_petsitter;
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.os.SystemClock
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
-import com.google.android.material.transition.MaterialSharedAxis
-import kr.co.lion.mungnolza.R
-import kr.co.lion.mungnolza.databinding.ActivityMatchingPetsitterBinding
-import kr.co.lion.mungnolza.ui.matching_petsitter.fragment.MatchingFragment
-import kr.co.lion.mungnolza.ui.matching_petsitter.fragment.PaymentFragment
-import kr.co.lion.mungnolza.ui.matching_petsitter.fragment.PetSitterInfoFragment
-import kr.co.lion.mungnolza.ui.matching_petsitter.fragment.PetSitterReviewFragment
-import kr.co.lion.mungnolza.ui.matching_petsitter.fragment.ReservationConfirmedFragment
-import kr.co.lion.mungnolza.util.MatchingPetsitterFragmentName
-
-class MatchingPetsitterActivity : AppCompatActivity() {
-
-    lateinit var activityMatchingPetsitterBinding: ActivityMatchingPetsitterBinding
-
-    // 프래그먼트의 주소값을 담을 프로퍼티
-    var oldFragment: Fragment? = null
-    var newFragment: Fragment? = null
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        activityMatchingPetsitterBinding = ActivityMatchingPetsitterBinding.inflate(layoutInflater)
-        setContentView(activityMatchingPetsitterBinding.root)
-
-        // 여기는 test를 위한 변경 공간입니다.
-
-        replaceFragment(MatchingPetsitterFragmentName.MATCHING_FRAGMENT,true,false,null)
-
-        // ReservationListActivity 실행한다.
-        /*val contentIntent = Intent(this, ReservationListActivity::class.java)
-        startActivity(contentIntent)
-        // MainActivity를 종료한다.
-        this.finish()*/
-
-        // RealtimeLocationActivity 실행한다.
-        /*val contentIntent = Intent(this, RealtimeLocationActivity::class.java)
-        startActivity(contentIntent)
-        // MainActivity를 종료한다.
-        this.finish()*/
-    }
-
-    // 지정한 Fragment를 보여주는 메서드
-    // name : 프래그먼트 이름
-    // addToBackStack : BackStack에 포함 시킬 것인지
-    // isAnimate : 애니메이션을 보여줄 것인지
-    // data : 새로운 프래그먼트에 전달할 값이 담겨져 있는 Bundle 객체
-    fun replaceFragment(name: MatchingPetsitterFragmentName, addToBackStack:Boolean, isAnimate:Boolean, data:Bundle?){
-
-        SystemClock.sleep(200)
-
-        // Fragment를 교체할 수 있는 객체를 추출한다.
-        val fragmentTransaction = supportFragmentManager.beginTransaction()
-
-        // oldFragment에 newFragment가 가지고 있는 Fragment 객체를 담아준다.
-        if(newFragment != null){
-            oldFragment = newFragment
-        }
-
-        // 이름으로 분기한다.
-        // Fragment의 객체를 생성하여 변수에 담아준다.
-        when(name){
-            // 매칭화면
-            MatchingPetsitterFragmentName.MATCHING_FRAGMENT -> {
-                newFragment = MatchingFragment()
-            }
-            // 펫시터 프로필 화면
-            MatchingPetsitterFragmentName.PETSITTER_INFO_FRAGMENT -> {
-                newFragment = PetSitterInfoFragment()
-            }
-            // 펫시터 리뷰 목록 화면
-            MatchingPetsitterFragmentName.PETSITTER_REVIEW_FRAGMENT -> {
-                newFragment = PetSitterReviewFragment()
-            }
-            // 결제 화면
-            MatchingPetsitterFragmentName.PAYMENT_FRAGMENT -> {
-                newFragment = PaymentFragment()
-            }
-            // 예약 확정 화면
-            MatchingPetsitterFragmentName.RESERVATION_CONFIRMED_FRAGMENT -> {
-                newFragment = ReservationConfirmedFragment()
-            }
-        }
-
-        // 새로운 Fragment에 전달할 객체가 있다면 arguments 프로퍼티에 넣어준다.
-        if(data != null){
-            newFragment?.arguments = data
-        }
-
-        if(newFragment != null){
-
-            // 애니메이션 설정
-            if(isAnimate == true){
-                // oldFragment -> newFragment
-                // oldFragment : exitTransition
-                // newFragment : enterTransition
-
-                // newFragment -> oldFragment
-                // oldFragment : reenterTransition
-                // newFragment : returnTransition
-
-                // MaterialSharedAxis : 좌우, 위아래, 공중 바닥 사이로 이동하는 애니메이션 효과
-                // X - 좌우
-                // Y - 위아래
-                // Z - 공중 바닥
-                // 두 번째 매개변수 : 새로운 화면이 나타나는 것인지 여부를 설정해준다.
-                // true : 새로운 화면이 나타나는 애니메이션이 동작한다.
-                // false : 이전으로 되돌아가는 애니메이션이 동작한다.
-
-                if(oldFragment != null){
-                    // old에서 new가 새롭게 보여질 때 old의 애니메이션
-                    oldFragment?.exitTransition = MaterialSharedAxis(MaterialSharedAxis.X, true)
-                    // new에서 old로 되돌아갈때 old의 애니메이션
-                    oldFragment?.reenterTransition = MaterialSharedAxis(MaterialSharedAxis.X, false)
-
-                    oldFragment?.enterTransition = null
-                    oldFragment?.returnTransition = null
-                }
-
-                // old에서 new가 새롭게 보여질 때 new의 애니메이션
-                newFragment?.enterTransition = MaterialSharedAxis(MaterialSharedAxis.X, true)
-                // new에서 old로 되돌아갈때 new의 애니메이션
-                newFragment?.returnTransition = MaterialSharedAxis(MaterialSharedAxis.X, false)
-
-                newFragment?.exitTransition = null
-                newFragment?.reenterTransition = null
-            }
-
-            // Fragment를 교체한다.(이전 Fragment가 없으면 새롭게 추가하는 역할을 수행한다)
-            // 첫 번째 매개 변수 : Fragment를 배치할 FragmentContainerView의 ID
-            // 두 번째 매개 변수 : 보여주고하는 Fragment 객체를
-            fragmentTransaction.replace(R.id.matchingPetsitterContainer, newFragment!!)
-
-            // addToBackStack 변수의 값이 true면 새롭게 보여질 Fragment를 BackStack에 포함시켜 준다.
-            if(addToBackStack == true){
-                // BackStack 포함 시킬때 이름을 지정해주면 원하는 Fragment를 BackStack에서 제거할 수 있다.
-                fragmentTransaction.addToBackStack(name.str)
-            }
-            // Fragment 교체를 확정한다.
-            fragmentTransaction.commit()
-        }
-    }
-
-    // BackStack에서 Fragment를 제거한다.
-    fun removeFragment(name: MatchingPetsitterFragmentName){
-        SystemClock.sleep(200)
-
-        // 지정한 이름으로 있는 Fragment를 BackStack에서 제거한다.
-        supportFragmentManager.popBackStack(name.str, FragmentManager.POP_BACK_STACK_INCLUSIVE)
-    }
-}
