--- MungNolZa/app/src/main/res/layout/activity_matching_petsitter.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:transitionGroup="true"
-    tools:context=".ui.matching_petsitter.MatchingPetsitterActivity">
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/matchingPetsitterContainer"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
-</LinearLayout>
